+++ model/nr-mac-scheduler-ue-info-pf.h
@@ -31,6 +31,7 @@ namespace ns3 {
  * to sort the UEs in case of a PF scheduler.
  *
  * \see CompareUeWeightsDl
+ * \see CompareUeWeightsUl
  */
 class NrMacSchedulerUeInfoPF : public NrMacSchedulerUeInfo
 {
@@ -48,7 +49,7 @@ public:
   }
 
   /**
-   * \brief Reset PF scheduler info
+   * \brief Reset DL PF scheduler info
    *
    * Set the last average throughput to the current average throughput,
    * and zeroes the average throughput as well as the current throughput.
@@ -60,12 +61,29 @@ public:
     m_lastAvgTputDl = m_avgTputDl;
     m_avgTputDl = 0.0;
     m_currTputDl = 0.0;
-    m_potentialTput = 0.0;
+    m_potentialTputDl = 0.0;
     NrMacSchedulerUeInfo::ResetDlSchedInfo ();
   }
 
   /**
-   * \brief Reset the avg Th to the last value
+   * \brief Reset UL PF scheduler info
+   *
+   * Set the last average throughput to the current average throughput,
+   * and zeroes the average throughput as well as the current throughput.
+   *
+   * It also calls NrMacSchedulerUeInfo::ResetUlSchedInfo.
+   */
+  virtual void ResetUlSchedInfo () override
+  {
+    m_lastAvgTputUl = m_avgTputUl;
+    m_avgTputUl = 0.0;
+    m_currTputUl = 0.0;
+    m_potentialTputUl = 0.0;
+    NrMacSchedulerUeInfo::ResetUlSchedInfo ();
+  }
+
+  /**
+   * \brief Reset the DL avg Th to the last value
    */
   virtual void ResetDlMetric () override
   {
@@ -74,21 +92,58 @@ public:
   }
 
   /**
+   * \brief Reset the UL avg Th to the last value
+   */
+  virtual void ResetUlMetric () override
+  {
+    NrMacSchedulerUeInfo::ResetUlMetric ();
+    m_avgTputUl = m_lastAvgTputUl;
+  }
+
+  /**
    * \brief Update the PF metric for downlink
    * \param totAssigned the resources assigned
    * \param timeWindow the time window
    * \param amc a pointer to the AMC
+   *
+   * Updates m_currTputDl and m_avgTputDl by keeping in consideration
+   * the assigned resources (in form of TBS) and the time window.
+   * It gets the tbSise by calling NrMacSchedulerUeInfo::UpdateDlMetric.
    */
   void UpdateDlPFMetric (const NrMacSchedulerNs3::FTResources &totAssigned,
                          double timeWindow,
                          const Ptr<const NrAmc> &amc);
+
   /**
-   * \brief Calculate the Potential throughput
+   * \brief Update the PF metric for uplink
+   * \param totAssigned the resources assigned
+   * \param timeWindow the time window
+   * \param amc a pointer to the AMC
+   *
+   * Updates m_currTputUl and m_avgTputUl by keeping in consideration
+   * the assigned resources (in form of TBS) and the time window.
+   * It gets the tbSise by calling NrMacSchedulerUeInfo::UpdateUlMetric.
+   */
+  void UpdateUlPFMetric (const NrMacSchedulerNs3::FTResources &totAssigned,
+                         double timeWindow,
+                         const Ptr<const NrAmc> &amc);
+
+  /**
+   * \brief Calculate the Potential throughput for downlink
    * \param assignableInIteration resources assignable
    * \param amc a pointer to the AMC
    */
-  void CalculatePotentialTPut (const NrMacSchedulerNs3::FTResources &assignableInIteration,
+  void CalculatePotentialTPutDl (const NrMacSchedulerNs3::FTResources &assignableInIteration,
                                const Ptr<const NrAmc> &amc);
+
+  /**
+   * \brief Calculate the Potential throughput for uplink
+   * \param assignableInIteration resources assignable
+   * \param amc a pointer to the AMC
+   */
+  void CalculatePotentialTPutUl (const NrMacSchedulerNs3::FTResources &assignableInIteration,
+                               const Ptr<const NrAmc> &amc);
+
 
   /**
    * \brief comparison function object (i.e. an object that satisfies the
@@ -96,7 +151,7 @@ public:
    * than (i.e. is ordered before) the second.
    * \param lue Left UE
    * \param rue Right UE
-   * \return true if the PF metric of the left UE are higher than the right UE
+   * \return true if the PF metric of the left UE is higher than the right UE
    *
    * The PF metric is calculated as following:
    *
@@ -111,17 +166,50 @@ public:
     auto luePtr = dynamic_cast<NrMacSchedulerUeInfoPF*> (lue.first.get ());
     auto ruePtr = dynamic_cast<NrMacSchedulerUeInfoPF*> (rue.first.get ());
 
-    double lPfMetric = std::pow (luePtr->m_potentialTput, luePtr->m_alpha) / std::max (1E-9, luePtr->m_avgTputDl);
-    double rPfMetric = std::pow (ruePtr->m_potentialTput, ruePtr->m_alpha) / std::max (1E-9, ruePtr->m_avgTputDl);
+    double lPfMetric = std::pow (luePtr->m_potentialTputDl, luePtr->m_alpha) / std::max (1E-9, luePtr->m_avgTputDl);
+    double rPfMetric = std::pow (ruePtr->m_potentialTputDl, ruePtr->m_alpha) / std::max (1E-9, ruePtr->m_avgTputDl);
 
     return (lPfMetric > rPfMetric);
   }
 
-  double m_currTputDl {0.0};    //!< Current slot throughput
-  double m_avgTputDl  {0.0};    //!< Average throughput during all the slots
-  double m_lastAvgTputDl {0.0}; //!< Last average throughput
-  double m_potentialTput {0.0}; //!< Potential throughput in one assignable resource (can be a symbol or a RBG)
+  /**
+   * \brief comparison function object (i.e. an object that satisfies the
+   * requirements of Compare) which returns ​true if the first argument is less
+   * than (i.e. is ordered before) the second.
+   * \param lue Left UE
+   * \param rue Right UE
+   * \return true if the PF metric of the left UE is higher than the right UE
+   *
+   * The PF metric is calculated as following:
+   *
+   * \f$ pfMetric_{i} = std::pow(potentialTPut_{i}, alpha) / std::max (1E-9, m_avgTput_{i}) \f$
+   *
+   * Alpha is a fairness metric. Please note that the throughput is calculated
+   * in bit/symbol.
+   */
+  static bool CompareUeWeightsUl (const NrMacSchedulerNs3::UePtrAndBufferReq &lue,
+                                  const NrMacSchedulerNs3::UePtrAndBufferReq &rue)
+  {
+    auto luePtr = dynamic_cast<NrMacSchedulerUeInfoPF*> (lue.first.get ());
+    auto ruePtr = dynamic_cast<NrMacSchedulerUeInfoPF*> (rue.first.get ());
+
+    double lPfMetric = std::pow (luePtr->m_potentialTputUl, luePtr->m_alpha) / std::max (1E-9, luePtr->m_avgTputUl);
+    double rPfMetric = std::pow (ruePtr->m_potentialTputUl, ruePtr->m_alpha) / std::max (1E-9, ruePtr->m_avgTputUl);
+
+    return (lPfMetric > rPfMetric);
+  }
+
+  double m_currTputDl {0.0};    //!< Current slot throughput in downlink
+  double m_avgTputDl  {0.0};    //!< Average throughput in downlink during all the slots
+  double m_lastAvgTputDl {0.0}; //!< Last average throughput in downlink
+  double m_potentialTputDl {0.0}; //!< Potential throughput in downlink in one assignable resource (can be a symbol or a RBG)
   float  m_alpha {0.0};         //!< PF fairness metric
+
+  double m_currTputUl {0.0};    //!< Current slot throughput in uplink
+  double m_avgTputUl  {0.0};    //!< Average throughput in uplink during all the slots
+  double m_lastAvgTputUl {0.0}; //!< Last average throughput in uplink
+  double m_potentialTputUl {0.0}; //!< Potential throughput in uplink in one assignable resource (can be a symbol or a RBG)
 };
+
 
 } // namespace ns3