+++ model/nr-spectrum-phy.h
@@ -248,6 +248,11 @@ public:
   void AddSrsSinrChunkProcessor (const Ptr<LteChunkProcessor>& p);
 
   /**
+   * \brief Adds the chunk processor that will process the received power
+   * \param p the chunk processor
+   */
+  void AddRsPowerChunkProcessor (const Ptr<LteChunkProcessor>& p);
+  /**
    * \brief SpectrumPhy that will be called when the SINR for the received
    * DATA is being calculated by the interference object over DATA chunk
    * processor