+++ model/nr-ue-net-device.cc
@@ -29,6 +29,7 @@
 #include <ns3/ipv4-l3-protocol.h>
 #include <ns3/object-map.h>
 #include <ns3/pointer.h>
+#include <ns3/ipv6-l3-protocol.h>
 
 namespace ns3 {
 
@@ -208,12 +209,10 @@ NrUeNetDevice::UpdateConfig (void)
 bool
 NrUeNetDevice::DoSend (Ptr<Packet> packet, const Address& dest, uint16_t protocolNumber)
 {
-  NS_LOG_FUNCTION (this << dest << protocolNumber);
-  if (protocolNumber != Ipv4L3Protocol::PROT_NUMBER)
-    {
-      NS_LOG_INFO ("unsupported protocol " << protocolNumber << ", only IPv4 is supported");
-      return false;
-    }
+  NS_LOG_FUNCTION (this << packet << dest << protocolNumber);
+  NS_ABORT_MSG_IF (protocolNumber != Ipv4L3Protocol::PROT_NUMBER
+                   && protocolNumber != Ipv6L3Protocol::PROT_NUMBER,
+                   "unsupported protocol " << protocolNumber << ", only IPv4 and IPv6 are supported");
   return m_nas->Send (packet, protocolNumber);
 }